+++ src/evpn_es_drain.rs
@@ -832,7 +832,7 @@ mod tests {
                         RibUpdate::SubscribeEvpnRouteEvents { reply } => {
                             let _ = reply.send(events_tx.subscribe());
                         }
-                        RibUpdate::QueryEvpnRoutes { reply } => {
+                        RibUpdate::QueryEvpnRoutes { reply, .. } => {
                             let _ = reply.send(Vec::new());
                         }
                         RibUpdate::InjectEvpn { route, reply } => {